examples/06_netauto_ecosystem/05_ios_upgrade_project/03_python_orchestrator_pluggable_cli_transfer_workflow/requirements.txt
netmiko>=4.4.0
scrapli>=2024.0.0
PyYAML>=6.0.1
